DX-997: Make cursor field in scan commands string (#1115)
* fix: make return type of cursor commands string

scan commands can return a '0' or a long number as a string. If we deserialize, we get 0 from '0' and a truncated number from the long number string. Second one is non-ideal.

In this case, we want to return the cursor as a string. This means excluding the cursor from the deserialization, hence the new deserializeScanResponse method.

* fix: reorder deserialize and commandOptions

this way, overwriting deserialize will be possible.